third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-flexbox/column-reverse-gap-expected.txt

This is a testharness.js-based test.
FAIL .flexbox 1 assert_equals: 
<div class="flexbox" data-expected-height="40">
  <div data-offset-y="30"></div>
  <div data-offset-y="15"></div>
  <div data-offset-y="0"></div>
</div>
offsetTop expected 15 but got 20
Harness: the test ran to completion.
